requirements.txt
anthropic
python-dotenv
rich
requests
nltk
pylint
autopep8
beautifulsoup4
Pillow
pydantic
tavily-python
transformers
